refactor instructions screen
