+++ src/webapp/components/DeliveryDialog/Instructions.js
@@ -1,21 +1,7 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import { makeStyles } from "@material-ui/core/styles";
 import Typography from "@material-ui/core/Typography";
-import DialogTitle from "@material-ui/core/DialogTitle";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogActions from "@material-ui/core/DialogActions";
-import Button from "@material-ui/core/Button";
-import instrutions from "./instructions.json";
-
-const useStyles = makeStyles(() => ({
-  moreInfoActionsContainer: {
-    justifyContent: "flex-start",
-  },
-  backButton: {
-    backgroundColor: "lightgray",
-  },
-}));
+import instructions from "./instructions.json";
 
 const Subtitle = ({ content }) => (
   <Typography variant="subtitle1">{content}</Typography>
@@ -67,7 +53,7 @@ const Instructions = () => {
 
   return (
     <>
-      {instrutions.map((block) => {
+      {instructions.map((block) => {
         const ComponentBlock = componentTypeMap[block.typename];
         const opts = {
           ...block,
@@ -82,29 +68,4 @@ const Instructions = () => {
   );
 };
 
-const InstructionsStep = ({ handleGoBack }) => {
-  const classes = useStyles();
-  const { t: str } = useTranslation();
-
-  return (
-    <>
-      <DialogTitle>
-        {str("webapp:deliveryNeeded.dialog.title", {
-          defaultValue: "How to make a delivery",
-        })}
-      </DialogTitle>
-      <DialogContent>
-        <Instructions />
-      </DialogContent>
-      <DialogActions className={classes.moreInfoActionsContainer}>
-        <Button onClick={handleGoBack} className={classes.backButton}>
-          {str("webapp:deliveryNeeded.dialog.backButton", {
-            defaultValue: "Back",
-          })}
-        </Button>
-      </DialogActions>
-    </>
-  );
-};
-
-export default InstructionsStep;
+export default Instructions;
